+++ src/main/java/global/Index.java
@@ -6,15 +6,15 @@ package global;
  */
 public class Index {
 
-    public static final int REF = 0;
-    public static final int HET = 1;
-    public static final int HOM = 2;
-    public static final int REF_MALE = 3;
-    public static final int HOM_MALE = 4;
-    public static final int MISSING = 5;
-    public static final int CTRL = 0;
-    public static final int CASE = 1;
-    public static final int ALL = 2;
-    public static final int GENOME = 0;
-    public static final int EXOME = 1;
+    public static final byte REF = 0;
+    public static final byte HET = 1;
+    public static final byte HOM = 2;
+    public static final byte REF_MALE = 3;
+    public static final byte HOM_MALE = 4;
+    public static final byte MISSING = 5;
+    public static final byte CTRL = 0;
+    public static final byte CASE = 1;
+    public static final byte ALL = 2;
+    public static final byte GENOME = 0;
+    public static final byte EXOME = 1;
 }
